+++ packages/snap-controller/src/Autocomplete/AutocompleteController.ts
@@ -998,7 +998,7 @@ export class AutocompleteController extends AbstractController {
 			this.track.product.addToCart(product);
 		});
 		if (products.length > 0) {
-			this.eventManager.fire('addToCart', { controller: this, products });
+			await this.eventManager.fire('addToCart', { controller: this, products });
 		}
 	};
 }