+++ packages/snap-preact/components/src/themes/snappy/components/organisms/index.ts
@@ -4,7 +4,6 @@ import { ThemeResponsiveComplete } from '../../../../providers';
 import { facet } from './facet';
 import { facetsHorizontal } from './facetsHorizontal';
 import { filterSummary } from './filterSummary';
-import { mobileSidebar } from './mobileSidebar';
 import { noResults } from './noResults';
 import { sidebar } from './sidebar';
 import { termsList } from './termsList';
@@ -17,7 +16,6 @@ export const organisms = {
 		...facet.default,
 		...facetsHorizontal.default,
 		...filterSummary.default,
-		...mobileSidebar.default,
 		...noResults.default,
 		...sidebar.default,
 		...toolbar.default,
@@ -28,7 +26,6 @@ export const organisms = {
 		...facet.mobile,
 		...facetsHorizontal.mobile,
 		...filterSummary.mobile,
-		...mobileSidebar.mobile,
 		...noResults.mobile,
 		...sidebar.mobile,
 		...toolbar.mobile,
@@ -39,7 +36,6 @@ export const organisms = {
 		...facet.tablet,
 		...facetsHorizontal.tablet,
 		...filterSummary.tablet,
-		...mobileSidebar.tablet,
 		...noResults.tablet,
 		...sidebar.tablet,
 		...toolbar.tablet,
@@ -50,7 +46,6 @@ export const organisms = {
 		...facet.desktop,
 		...facetsHorizontal.desktop,
 		...filterSummary.desktop,
-		...mobileSidebar.desktop,
 		...noResults.desktop,
 		...sidebar.desktop,
 		...toolbar.desktop,