+++ packages/snap-preact/src/Templates/Stores/TemplateStore.ts
@@ -29,6 +29,7 @@ import type {
 import type {
 	PluginAddToCartConfig,
 	PluginBackgroundFiltersConfig,
+	PluginKlaviyoEventsConfig,
 	PluginLoggerConfig,
 	PluginScrollToTopConfig,
 } from '@athoscommerce/snap-platforms/common';
@@ -47,7 +48,7 @@ import type {
 import type { GlobalThemeStyleScript, IntegrationPlatforms } from '../../types';
 import type { ClientConfig } from '@athoscommerce/snap-client';
 import { RecommendationInstantiatorConfigSettings } from '../../Instantiators/RecommendationInstantiator';
-
+import type { PluginMarketsConfig } from '@athoscommerce/snap-platforms/shopify';
 export type TemplateThemeTypes = 'library' | 'local';
 export type TemplateTypes = 'search' | 'autocomplete' | `recommendation/${RecsTemplateTypes}`;
 
@@ -104,6 +105,7 @@ type WindowProperties = {
 
 type TemplatesStoreThemeConfigLocked = {
 	extends: keyof LibraryImports['theme'];
+	globalResultComponent?: string;
 	style?: GlobalThemeStyleScript;
 	variables?: ThemeVariablesPartial;
 	overrides?: ThemeResponsiveComplete;
@@ -133,11 +135,13 @@ export type CommonPlugins = {
 	scrollToTop?: PluginScrollToTopConfig;
 	logger?: PluginLoggerConfig;
 	addToCart?: PluginAddToCartConfig;
+	klaviyoEvents?: PluginKlaviyoEventsConfig;
 };
 export type ShopifyPlugins = {
 	backgroundFilters?: PluginShopifyBackgroundFiltersConfig;
 	mutateResults?: PluginShopifyMutateResultsConfig;
 	addToCart?: PluginShopifyAddToCartConfig;
+	markets?: PluginMarketsConfig;
 };
 
 export type BigCommercePlugins = {